+++ CHANGELOG.md
@@ -1,3 +1,7 @@
+1.10.18:
+    - Added ColorSchemeAnimator to animate color scheme changing from one color scheme to another.
+    - Added public ColorSchemeComponent methods required to create custom animators.
+
 1.10.17:
     - Fixed an issue with InfoBox, that can cause build errors.